moving existing diagnostics into new folder
